:fire: remove repo file system openner :tractor: relocate old code to deprecated
