+++ docs/level-10-moban-dependency-as-git-repo/.moban.yml
@@ -3,7 +3,7 @@ configuration:
     - "git://github.com/moremoban/pypi-mobans.git!/templates"
     - local
   configuration: config.yml
-  configuration_dir: "repo://pypi-mobans/config"
+  configuration_dir: "git://github.com/moremoban/pypi-mobans.git!/config"
 targets:
   - mytravis.yml: travis.yml.jj2
   - test.txt: demo.txt.jj2
